requirements.txt
# Road Comfort System - Python Dependencies

# Web Framework & APIs
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning & Deep Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.14.0
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Data Processing
opencv-python==4.8.1.78
librosa==0.10.0

# Database & Caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
motor==3.3.2

# GIS & Map Matching
geopy==2.4.0
shapely==2.0.2
geopandas==0.14.1
osmnx==1.9.2

# Config & Logging
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
ipython==8.18.1
jupyter==1.0.0

# Model Export & Deployment
onnx==1.15.0
onnxruntime==1.17.0
joblib==1.3.2

# Utilities
requests==2.31.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
